Remove Workspace state dependency from driver
This modification reduces the dependency of the cmake driver layer to the  workspace context. This modification allows to setup low level CMake driver tests, so that the driver code could be tested with different cmake versions.
The UI parts can then tested with a driver mock, or by system tests.

- Replace DirectoryContext by ConfigurationReader
- Moved access to DirectoryContext state by setter funcions (setVarianOptions). This allows to set variant in a consistent way.
- Reduces fakes from tests