Fix problem with async setup and clang in cygwin test
The setup function of the cmake server driver was executed async without call of done.
In some cases (appveyor) it failed.

The cygwin test tried different compilers, this modification removes the clang compiler from
cygwin. The tests should only test that cygwin principle works, not the test of all compilers.